feat(service): 优化商品列表和订单列表的查询逻辑
- 修改商品列表查询接口，使用 Long 类型替代 Double 类型表示分数- 优化订单列表查询逻辑，使用 Redis 缓存提高性能- 调整产品和服务接口，以适应新的查询逻辑
- 修复了一些与数据类型和查询结果相关的小问题
- 修理一些事务传播属性以更符合其所在场景

+++ parent/service/src/main/java/com/zhuxi/Result/ProductPageResult.java
@@ -11,13 +11,13 @@ public class ProductPageResult<T>{
     @Schema(description = "当前页数据")
     private List<T> items;
     @Schema(description = "下一页游标(redis缓存使用)")
-    private Double nextCursor;
+    private Long nextCursor;
     @Schema(description = "是否有下一页(当为数据库查询时，其为null,若正常显示则会查缓存)")
     private boolean hasNext;
     @Schema(description = "本页最后一条数据的id(用于未命中时兜底)",requiredMode = Schema.RequiredMode.REQUIRED)
     private Long lastId;
 
-    public ProductPageResult(List<T> items, Double nextCursor, boolean hasNext,Long lastId) {
+    public ProductPageResult(List<T> items, Long nextCursor, boolean hasNext,Long lastId) {
         this.items = items;
         this.nextCursor = nextCursor;
         this.hasNext = hasNext;
@@ -32,11 +32,11 @@ public class ProductPageResult<T>{
         this.items = items;
     }
 
-    public Double getNextCursor() {
+    public Long getNextCursor() {
         return nextCursor;
     }
 
-    public void setNextCursor(Double nextCursor) {
+    public void setNextCursor(Long nextCursor) {
         this.nextCursor = nextCursor;
     }
